# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a PoC playground rest server",
        "title": "Playground REST server",
        "termsOfService": "http://codenotary.com/terms",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/exec/python": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "info"
                ],
                "summary": "Execute a python script",
                "operationId": "pythonExec",
                "parameters": [
                    {
                        "description": "Run request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.runRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.runResponse"
                        }
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
        "main.OutputLine": {
            "type": "object",
            "properties": {
                "flux": {
                    "type": "string"
                },
                "line": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "number"
                }
            }
        },
...
        "main.runRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "immudb": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "main.runResponse": {
            "type": "object",
            "properties": {
                "immudb": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "output": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.OutputLine"
                    }
                },
                "tree": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    }
}